feat(query-builder): save and restore inner query state

+++ projects/igniteui-angular/src/lib/query-builder/query-builder-tree.component.ts
@@ -219,7 +219,7 @@ export class IgxQueryBuilderTreeComponent implements AfterViewInit, OnDestroy {
     * Returns the expression tree.
     */
     public get expressionTree(): IExpressionTree {
-        return this._expressionTree;
+        return this._newEntityExpressionTree ?? this._expressionTree;
     }
 
     /**
@@ -227,6 +227,10 @@ export class IgxQueryBuilderTreeComponent implements AfterViewInit, OnDestroy {
      */
     @Input()
     public set expressionTree(expressionTree: IExpressionTree) {
+        if (this._newEntityExpressionTree) {
+            this._newEntityExpressionTree = null;
+        }
+
         if (JSON.stringify(expressionTree) !== JSON.stringify(this._expressionTree)) {
             this._expressionTree = expressionTree;
             if (!expressionTree) {
@@ -467,7 +471,6 @@ export class IgxQueryBuilderTreeComponent implements AfterViewInit, OnDestroy {
 
     private destroy$ = new Subject<any>();
     private _parentExpression: ExpressionOperandItem;
-    private _initialExpressionTree: IExpressionTree;
     private _selectedEntity: EntityType;
     private _selectedReturnFields: string | string[];
     private _selectedField: FieldType;
@@ -486,6 +489,8 @@ export class IgxQueryBuilderTreeComponent implements AfterViewInit, OnDestroy {
     private _locale;
     private _entityNewValue: EntityType;
     private _resourceStrings = getCurrentResourceStrings(QueryBuilderResourceStringsEN);
+    private _newEntityExpressionTree: IExpressionTree = null;
+    private _initialState: IExpressionTree = null;
 
     public get level(): number {
         let parent = this.elRef.nativeElement.parentElement;
@@ -543,6 +548,13 @@ export class IgxQueryBuilderTreeComponent implements AfterViewInit, OnDestroy {
     /**
      * @hidden @internal
      */
+    public set selectedEntity(value: string) {
+        this._selectedEntity = this.entities?.find(el => el.name === value);
+    }
+
+    /**
+     * @hidden @internal
+     */
     public get selectedEntity(): EntityType {
         return this._selectedEntity;
     }
@@ -580,21 +592,27 @@ export class IgxQueryBuilderTreeComponent implements AfterViewInit, OnDestroy {
      * @hidden
      */
     public onEntityChangeConfirm() {
+        if (this._parentExpression) {
+            this._newEntityExpressionTree = this.createExpressionTreeFromGroupItem(this.createExpressionGroupItem(this._expressionTree));
+        }
+        
         this._selectedEntity = this._entityNewValue;
         if (!this._selectedEntity.fields) {
             this._selectedEntity.fields = [];
         }
         this.fields = this._entityNewValue ? this._entityNewValue.fields : [];
-
         this._selectedReturnFields = this._entityNewValue.fields?.map(f => f.field);
 
-        if (this._expressionTree) {
-            this._expressionTree.entity = this._selectedEntity.name;
-            this._expressionTree.returnFields = [];
-            this._expressionTree.filteringOperands = [];
+        const exprTree = this._newEntityExpressionTree ?? this._expressionTree;
+        if (exprTree) {
+            exprTree.entity = this._entityNewValue.name;
+            exprTree.returnFields = [];
+            exprTree.filteringOperands = [];
 
             this._editedExpression = null;
-            this.expressionTreeChange.emit(this._expressionTree);
+            if (!this.parentExpression) {
+                this.expressionTreeChange.emit(exprTree);
+            }
 
             this.addAndGroup();
         }
@@ -618,7 +636,7 @@ export class IgxQueryBuilderTreeComponent implements AfterViewInit, OnDestroy {
         if (this._selectedReturnFields !== value && oldValue !== value) {
             this._selectedReturnFields = value;
 
-            if (this._expressionTree) {
+            if (this._expressionTree && !this.parentExpression) {
                 this._expressionTree.returnFields = value;
                 this.expressionTreeChange.emit(this._expressionTree);
             }
@@ -767,6 +785,7 @@ export class IgxQueryBuilderTreeComponent implements AfterViewInit, OnDestroy {
         }
 
         this._expressionTree = this.createExpressionTreeFromGroupItem(this.rootGroup, this.selectedEntity?.name, this.selectedReturnFields);
+        this._newEntityExpressionTree = null;
         this.expressionTreeChange.emit(this._expressionTree);
     }
 
@@ -787,23 +806,29 @@ export class IgxQueryBuilderTreeComponent implements AfterViewInit, OnDestroy {
         if (this.innerQueries) {
             const innerQuery = this.innerQueries.filter(q => q.isInEditMode())[0];
             if (innerQuery) {
-                innerQuery.expressionTree = this._initialExpressionTree;
-
+                
                 if (innerQuery._editedExpression) {
                     innerQuery.cancelOperandEdit();
                 }
+
+                innerQuery.expressionTree = this._initialState;
+                innerQuery.selectedEntity = this._initialState.entity;
+                innerQuery.selectedReturnFields = this._initialState.returnFields;
+
             }
         }
 
         if (this._editedExpression) {
             this._editedExpression.inEditMode = false;
 
-            if (!this._editedExpression.expression.fieldName) {
+            if (!this._editedExpression.expression.fieldName && !this._initialState) {
                 this.deleteItem(this._editedExpression);
             }
 
             this._editedExpression = null;
         }
+
+        this._initialState = null;
     }
 
     /**
@@ -811,6 +836,7 @@ export class IgxQueryBuilderTreeComponent implements AfterViewInit, OnDestroy {
      */
     public operandCanBeCommitted(): boolean {
         const innerQuery = this.innerQueries.filter(q => q.isInEditMode())[0];
+        
         return this.selectedField && this.selectedCondition &&
             (
                 (
@@ -818,7 +844,7 @@ export class IgxQueryBuilderTreeComponent implements AfterViewInit, OnDestroy {
                     !(this.selectedField?.filters?.condition(this.selectedCondition)?.isNestedQuery)
                 ) ||
                 (
-                    innerQuery && !!innerQuery.expressionTree && innerQuery._editedExpression == undefined && innerQuery.selectedReturnFields.length > 0
+                    innerQuery && !!innerQuery.expressionTree && innerQuery._editedExpression == undefined && innerQuery.expressionTree.filteringOperands.length > 0
                 ) ||
                 this.selectedField.filters.condition(this.selectedCondition).isUnary
             );
@@ -907,9 +933,17 @@ export class IgxQueryBuilderTreeComponent implements AfterViewInit, OnDestroy {
         expressionItem.inEditMode = true;
         this._editedExpression = expressionItem;
         if (expressionItem.expression.searchTree) {
-            this._initialExpressionTree = expressionItem.expression.searchTree;
+            this._initialState = 
+            {
+                filteringOperands: [],
+                operator: expressionItem.expression.searchTree.operator,
+                fieldName: expressionItem.expression.searchTree.fieldName,
+                entity: expressionItem.expression.searchTree.entity,
+                returnFields: expressionItem.expression.searchTree.returnFields
+            };
+            expressionItem.expression.searchTree.filteringOperands.forEach(o => this._initialState.filteringOperands.push(o));
         } else {
-            this._initialExpressionTree = null;
+            this._initialState = null;
         }
 
         this.cdr.detectChanges();
